+++ CHANGELOG.md
@@ -15,6 +15,8 @@
 
 ### Fixed
 - Fixed a bug where Neo block tabs weren't showing indicators of validation errors in descendant blocks, if the descendant blocks were hidden in child blocks UI element(s)
+- Fixed a bug where the block type names on existing Neo input blocks weren't being translated
+- Fixed a bug where the tab names on all Neo input blocks weren't being translated
 
 ## 3.3.9 - 2022-09-16